feat: more nuanced unauthorized and forbidden handling
When `x-rems-api-key` is used, even if it's wrong, let's consider it
as an attempt at using the API, and give the errors accordingly. For
example if the `x-rems-user-id` is missing, you are essentially
unauthorized.

Some of the tests even documented the slightly odd previous behavior
when we went to check the CSRF token while it was clearly an API call.